Moving exceptions and utils.
