+++ dces-drc-integration/src/main/java/uk/gov/justice/laa/crime/dces/integration/exception/DcesDrcServiceException.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.dces.integration.rest.common.exception;
+package uk.gov.justice.laa.crime.dces.integration.exception;
 
 public class DcesDrcServiceException extends RuntimeException {
     public DcesDrcServiceException(final String errorMessage, final Throwable originalException) {
